实战案例/案例5-多模态产品问答系统/requirements.txt
# 案例5依赖
streamlit==1.29.0
Pillow==10.2.0
numpy==1.26.3
